Project_3a/kevin_gazebo/requirements.txt
absl-py==0.12.0
actionlib==1.12.1
angles==1.9.12
apturl==0.5.2
argon2-cffi==20.1.0
asn1crypto==0.24.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
base-local-planner==1.16.7
bleach==1.5.0
bondpy==1.8.5
Brlapi==0.6.6
cachetools==4.2.1
camera-calibration==1.15.0
camera-calibration-parsers==1.11.13
catkin==0.7.29
catkin-pkg==0.4.23
certifi==2018.1.18
cffi==1.14.5
chardet==3.0.4
click==6.7
colorama==0.4.4
command-not-found==0.3
contextlib2==0.6.0.post1
controller-manager==0.18.3
controller-manager-msgs==0.18.3
cryptography==2.1.4
cupshelpers==1.0
cv-bridge==1.13.0
cycler==0.10.0
Cython==0.29.23
decorator==5.0.7
defer==1.0.6
defusedxml==0.7.1
diagnostic-analysis==1.9.7
diagnostic-common-diagnostics==1.9.7
diagnostic-updater==1.9.7
distro==1.5.0
distro-info===0.18ubuntu0.18.04.1
docutils==0.17.1
drone-wrapper==1.3.7
dynamic-reconfigure==1.6.3
dynamixel-sdk==3.7.31
entrypoints==0.3
enum34==1.1.10
flatbuffers==1.12
gast==0.3.3
gazebo-plugins==2.8.7
gazebo-ros==2.8.7
gencpp==0.6.5
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.16
gennodejs==2.0.1
genpy==0.6.14
google-auth==1.29.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
html5lib==0.9999999
httplib2==0.9.2
idna==2.6
image-geometry==1.13.0
importlib-metadata==3.10.1
interactive-markers==1.11.5
ipykernel==5.5.3
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
jeepney==0.6.0
Jinja2==2.11.3
joint-state-publisher==1.12.15
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kdl-parser-py==1.13.1
Keras-Preprocessing==1.1.2
keyring==23.0.1
keyrings.alt==3.0
kiwisolver==1.3.1
language-selector==0.1
laser-geometry==1.6.7
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
linecache2==1.0.0
louis==3.5.0
lxml==4.6.3
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.3.4
MarkupSafe==1.0
matplotlib==3.3.4
mavros==1.6.0
message-filters==1.14.10
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.10.4
notebook==6.3.0
numpy==1.19.5
oauth==1.0.1
oauthlib==3.1.0
olefile==0.45.1
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.2
pbr==3.1.1
pexpect==4.2.1
pickleshare==0.7.5
Pillow==7.0.0
pkginfo==1.7.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.15.8
ptyprocess==0.7.0
py-trees==0.6.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.8.1
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.7
pyRFC3339==1.0
pyrsistent==0.17.3
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.8.1
python-debian==0.1.32
python-qt-binding==0.4.3
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
pyzmq==22.0.3
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
qtconsole==5.0.3
QtPy==1.9.0
readme-renderer==29.0
reportlab==3.4.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests-unixsocket==0.1.5
resource-retriever==1.12.6
rfc3986==1.4.0
rosbag==1.14.10
rosboost-cfg==1.14.9
rosclean==1.14.9
roscreate==1.14.9
rosgraph==1.14.10
roslaunch==1.14.10
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.10
rosmake==1.14.9
rosmaster==1.14.10
rosmsg==1.14.10
rosnode==1.14.10
rosparam==1.14.10
rospkg==1.3.0
rospy==1.14.10
rosserial-arduino==0.8.0
rosserial-client==0.8.0
rosserial-python==0.8.0
rosservice==1.14.10
rostest==1.14.10
rostopic==1.14.10
rosunit==1.14.9
roswtf==1.14.10
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.9
rqt-controller-manager==0.18.3
rqt-dep==0.4.9
rqt-drone-teleop==1.3.7
rqt-ez-publisher==0.5.0
rqt-graph==0.4.11
rqt-ground-robot-teleop==1.3.7
rqt-gui==0.5.2
rqt-gui-py==0.5.2
rqt-image-view==0.4.16
rqt-joint-trajectory-controller==0.17.2
rqt-joint-trajectory-plot==0.0.5
rqt-launch==0.4.8
rqt-launchtree==0.2.0
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-multiplot==0.0.10
rqt-nav-view==0.5.7
rqt-play-motion-builder==1.0.2
rqt-plot==0.4.13
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.2
rqt-py-console==0.4.8
rqt-py-trees==0.3.1
rqt-reconfigure==0.5.3
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.13
rqt-robot-steering==0.5.10
rqt-rotors==2.2.3
rqt-runtime-monitor==0.5.7
rqt-rviz==0.6.0
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.11
rqt-virtual-joy==0.1.2
rqt-web==0.4.8
rsa==4.7.2
rviz==1.13.16
scour==0.36
SecretStorage==3.3.1
Send2Trash==1.5.0
sensor-msgs==1.12.8
simplejson==3.13.2
six==1.15.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.5
system-service==0.3
systemd-python==234
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
tensorflow-object-detection-api==0.1.1
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
testresources==2.0.0
tf==1.12.1
tf-conversions==1.12.1
tf-object-detection-util==0.1.8
tf2-geometry-msgs==0.6.5
tf2-kdl==0.6.5
tf2-py==0.6.5
tf2-ros==0.6.5
topic-tools==1.14.10
tornado==6.1
tqdm==4.60.0
traceback2==1.4.0
traitlets==4.3.3
turtlebot3-example==1.2.5
turtlebot3-teleop==1.2.5
twine==3.4.1
typing-extensions==3.7.4.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unique-id==1.0.6
unittest2==1.1.0
urdfdom-py==0.4.5
urllib3==1.22
usb-creator==0.3.3
wadllib==1.3.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xacro==1.13.9
xkit==0.0.0
zipp==3.4.1
zope.interface==4.3.2
